try ... catch around xmlhttprequest.send() in httpHandler.
Why? Apparently if we try to access a resource in the .xpi that does not
exist, the request does not fail and run the onerror() function, as it
does when accessing proper URLs (including chrome:// URLs when
cliqz@cliqz.com is only a symbolic link in extensions/!!!), but throws
an NS_ERROR_FILE_NOT_FOUND right away.